Modularize the function into each dedicated module
- Aim of this change to maintain code, and make each module more cohesive.
